Added the session management logic.

+++ app/src/main/java/com/example/firstandroidapp/ShowDetails.java
@@ -1,5 +1,9 @@
 package com.example.firstandroidapp;
 
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
@@ -49,7 +53,7 @@ public class ShowDetails extends AppCompatActivity {
         emailView.setText(sharedPreferences.getString(Utility.emailAddressKey, ""));
         genderView.setText(sharedPreferences.getString(Utility.genderKey, ""));
 
-        if (sharedPreferences.getString(Utility.countryKey, "").equals("Select Country") )
+        if (sharedPreferences.getString(Utility.countryKey, "").equals("Select Country"))
             countryLL.setVisibility(View.GONE);
         else countryView.setText(sharedPreferences.getString(Utility.countryKey, ""));
 
@@ -67,12 +71,48 @@ public class ShowDetails extends AppCompatActivity {
         logoutButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                editor.clear();
-                editor.apply();
-                Toast.makeText(ShowDetails.this, "Log out Successfully", Toast.LENGTH_SHORT).show();
-                finish();
+                // Called custom alert dialog box.
+                customAlertDialogBoxCreation();
             }
         });
     }
+
+    private void customAlertDialogBoxCreation() {
+        // Create the object of AlertDialog Builder class
+        AlertDialog.Builder builder = new AlertDialog.Builder(ShowDetails.this);
+
+        // Set the message show for the Alert time
+        builder.setMessage(getString(R.string.dialog_box_message));
+
+        // Set Alert Title
+        builder.setTitle(getString(R.string.dialog_box_title));
+
+        // Set Cancelable false for when the user clicks on the outside the Dialog Box then it will remain show
+        builder.setCancelable(false);
+
+        // Set the positive button with yes name Lambda OnClickListener method is use of DialogInterface interface.
+        builder.setPositiveButton("Yes", (DialogInterface.OnClickListener) (dialog, which) -> {
+            // When the user click yes button the activity will be finished and the the values of the shared preference will be cleared.
+            editor.clear();
+            editor.apply();
+            // TO finish the activity
+            Intent i = new Intent(ShowDetails.this, LoginPage.class);
+            startActivity(i);
+            finish();
+        });
+
+        // Set the Negative button with No name Lambda OnClickListener method is use of DialogInterface interface.
+        builder.setNegativeButton("No", (DialogInterface.OnClickListener) (dialog, which) -> {
+            // If user click no then dialog box is canceled.
+            dialog.cancel();
+        });
+
+        // Create the Alert dialog
+        AlertDialog alertDialog = builder.create();
+        // Show the Alert Dialog box
+        alertDialog.show();
+
+    }
+
 
 }
